feat: adicionado Exceptions de Transação e Categoria, pesquisa detalhada, atualizações e exclusão de controllers que faltavam.

+++ src/main/java/io/github/financasapi/apifinancas/service/UsuarioService.java
@@ -1,5 +1,8 @@
 package io.github.financasapi.apifinancas.service;
 
+import io.github.financasapi.apifinancas.expections.OperacaoNaoPermitidaException;
+import io.github.financasapi.apifinancas.expections.UsuarioNaoEncontradoException;
+import io.github.financasapi.apifinancas.model.Categoria;
 import io.github.financasapi.apifinancas.model.Usuario;
 import io.github.financasapi.apifinancas.repository.UsuarioRepository;
 import io.github.financasapi.apifinancas.validador.UsuarioValidator;
@@ -16,9 +19,10 @@ public class UsuarioService {
 
     private final UsuarioRepository repository;
     private final UsuarioValidator validator;
+    private final UsuarioRepository usuarioRepository;
 
     public Usuario salvar(Usuario usuario) {
-        validator.validar(usuario);
+        validator.validarUser(usuario);
         return repository.save(usuario);
     }
 
@@ -34,6 +38,17 @@ public class UsuarioService {
     }
 
     public void deletar(UUID id) {
+        if(!repository.existsById(id)) {
+            throw new UsuarioNaoEncontradoException("Usuário com o ID informado não existe.");
+        }
         repository.deleteById(id);
+    }
+
+    public Usuario atualizar(UUID id, Usuario usuario) {
+        Usuario usuarioExiste = usuarioRepository.findById(id).orElseThrow(() -> new OperacaoNaoPermitidaException("Usuário não encontrado."));
+        usuarioExiste.setNome(usuario.getNome());
+        usuarioExiste.setEmail(usuario.getEmail());
+        usuarioExiste.setSenha(usuario.getSenha());
+        return usuarioRepository.save(usuarioExiste);
     }
 }
